refactor: Move Button component to global/ui
