+++ src/global/ui/GlobalButton.tsx
@@ -9,7 +9,7 @@ interface ButtonProps {
   variant?: 'default' | 'secondary';
   children: React.ReactNode;
   onClick?: () => void;
-  disabled?: boolean; // ✅ disabled 추가
+  disabled?: boolean; // disabled 추가
 }
 
 export function AuthButton({
@@ -20,9 +20,9 @@ export function AuthButton({
 }: ButtonProps) {
   return (
     <Button
-      className={`${variantClasses[variant]} w-full h-[55px] rounded-lg font-semiBold text-[16px] font-white hover:bg-primary-1`}
-      onClick={!disabled ? onClick : undefined} // ✅ disabled 상태에서는 클릭 이벤트 실행 X
-      disabled={disabled} // ✅ disabled 명확하게 적용
+      className={`${variantClasses[variant]} w-full h-[55px] rounded-lg font-semibold text-[16px] text-[#ffffff] hover:bg-primary-1`}
+      onClick={!disabled ? onClick : undefined} // disabled 상태에서는 클릭 이벤트 실행 X
+      disabled={disabled} // disabled 명확하게 적용
     >
       {children}
     </Button>
