+++ src/user/components/mentor/MentorSetup.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { AuthButton } from '@/auth/components/AuthButton';
+import { AuthButton } from '@/global/ui/GlobalButton';
 import FileUpload from '@/user/components/FileUpload';
 import MentorProfile from '@/user/components/mentor/MentorProfile'; // 프로필 설정 컴포넌트 추가
